LaserTower

+++ Assets/Scripts/Enemies/Enemy.cs
@@ -37,7 +37,7 @@ public class Enemy : MonoBehaviour
 
     [Header("Damage")]
     [SerializeField] public int lifePoints = 50;
-    private int currentHealth;
+    private float currentHealth;
     [SerializeField] public float nexusDamage = 10;
     [SerializeField] private float timeToExplode = 2;
     
@@ -98,14 +98,14 @@ public class Enemy : MonoBehaviour
 
     private IEnumerator StopAndApplyDamageToBase()
     {
-    yield return new WaitForSeconds(timeToExplode);
+        yield return new WaitForSeconds(timeToExplode);
         nbreMana = 0;
         Destroy(gameObject);
 
     
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         currentHealth -= amount;
         if (currentHealth <= 0)
